[STASH2] msg over RDMA ... almost there (I think)
NOTE TO SELF
============

samp -> agg1 ;  64K sets, 10 sec worked with
                the send/recv credit is 1, and CQ post credit is 128.

When credit is more than 1, sometimes we got garbage RECV message. I
think this is because the next message was not fully copied, but we just
looked into the header and thought that it was available to be processed
(I'm guessing that the lower-bytes were copied first). Will try putting
the check bits at the end later.