[STASH2] 64K sets samp - L1 - L2 (5 sec) works!
NOTE: After moving the `processed` status of the message buffer to the
end of the buffer, we don't see garbage recv message anymore. This has
been tested with samp - L1 - L2 with 5 sec interval with the following
number of sets: 64K, 32K, 16K, 8K, 4K, 1K, 128, 64.

With 1 sec interval, it is good up to 16K.

With 32K, 3 sec interval is good. But, 1 sec and 2 sec intervals got
INVALID_PARAM at one of the PostRdma().

Similarly, with 64K, but 3 sec interval didn't work. 5 sec works.