Refactor the lookup error path
The lookup error path will call ldms_set_delete which in turn
will call ldms_xprt_set_delete which informs downstream
aggregators of the deleted set. However, in the error path
the downstream peer has never received a dir_add because
the set was never published. This results in the downstream
reporting an error that a dir_delete was received for a
set that was never added.

It also fixes the case where the set is created in the local
tree, but an error occurs in the RDMA of the meta data into
the local set. In this case, the set was not being removed
from the set tree which would cause all subsequent lookup
requests to fail synchronously because the set was already
present in the tree.