Change munge authentication message
The munge authentication message previously consisted of
the local and remote addresses of the transport between
the two peers. This message presumed that both peers see
the same addresses (albeit swapped). This assumption is
false in many routed and NAT'd environments.

This patch changes the message to "MungeAuthentication"
which is validated by the receiving peer. In order to
support backward compatability with older plugins, an
option 'compat' has been added. This option, when set, will
use the previous message format so that newer aggregators
can connect to older samplers. The receive path supports
either, specifically, if the new message doesn't match
it tries the old format.